Seperate all these constants into one file for easy viewing and updating
