+++ constants.ts
@@ -1,0 +1,14 @@
+//Solana and token variables
+export const USE_MAINNET = false;
+export const VERIFIED_CREATOR = "AHx6cQKhJQ6vV9zhb37B7gKtRRGDuLtTfNWgvMiLDJp7";
+export const TOKEN_TO_SEND = "toKPe7ENJiRBANPLnnp4dHs7ccFyHRg2oSEPdDfumg8";
+export const AMOUNT_TO_SEND = 42000000000; // including decimals
+
+//Filename inputs and outputs
+export const KEYPAIR_FILE = "C:\\Users\\loopc\\wkdir\\DEVkasD4qwUJQ8LS7rcJHcGDQQzrEtWgk2jB6v5FHngo.json";
+export const PROGRESS_FILE_PATH = "./progress.json";
+export const ERROR_FILE_PATH = "./error.json";
+
+//secondary Market wallets
+export const MAGIC_EDEN_ADDRESS = "GUfCR9mK6azb9vcpsxgXyj7XRPAKJd4KMHTTVvtncGgp";
+export const MAGIC_EDEN_ESCROW = "1BWutmTvYPwDtmw9abTkS4Ssr8no61spGAvW1X6NDix";
