+++ prepareAirdrop.ts
@@ -5,31 +5,11 @@ import { assert } from 'console';
 import * as splToken from "@solana/spl-token";
 import { token } from '@project-serum/anchor/dist/cjs/utils';
 import * as fs from 'fs';
+import { TokenInfo } from './tokenInfo';
+import { PROGRESS_FILE_PATH, AMOUNT_TO_SEND, TOKEN_TO_SEND, VERIFIED_CREATOR } from './constants';
 
-const VERIFIED_CREATOR = "AHx6cQKhJQ6vV9zhb37B7gKtRRGDuLtTfNWgvMiLDJp7";
-const TOKEN_TO_SEND = "toKPe7ENJiRBANPLnnp4dHs7ccFyHRg2oSEPdDfumg8";
-const AMOUNT_TO_SEND = 42000000000; // including decimals
 
-export const PROGRESS_FILE_PATH = "./progress.json";
 
-export class TokenInfo {
-    metadataAccount: string;
-    nftTokenMint: string;
-    nftName: string | undefined;
-    ownerWallet: string | undefined;
-    sendableTokenMint: string | undefined;
-    sendableAmount: number | undefined;
-    txid: string | undefined;
-
-    constructor(metadataAccount: string, tokenMint: string) {
-        this.metadataAccount = metadataAccount;
-        this.nftTokenMint = tokenMint;
-    }
-
-    show(){
-        console.log(this.metadataAccount + " -> " + this.nftTokenMint +' -> '+ this.ownerWallet);
-    }
-}
 
 export function writeJson(data: TokenInfo[]){
     let json = JSON.stringify(data);
